Update Kotlin and AGP versions, address Room warning (#264)
This commit updates the Kotlin version to 2.0.0 and the Android Gradle Plugin (AGP) to 8.11.1.

It also addresses a Room `CURSOR_MISMATCH` warning in `PaymentIndicatorDao.kt` by adding the `@SuppressWarnings(RoomWarnings.CURSOR_MISMATCH)` annotation to the affected query.

Additionally, the `@Transaction` annotation was added to the `getById` function in `CompanyLocationDao.kt` to ensure atomicity when fetching related entities.

Signed-off-by: Comfort Mwalija <calmwalija@gmail.com>